refactor(protocol): move composerInput into the package
