+++ cmd/evener-hub/frontend/src/protocol/index.ts
@@ -43,6 +43,8 @@ export { translateAttachmentMarkers } from "./attachmentMarkers";
 export type { AnyNotification, AppwireClientOptions, ConnectionState, TerminalReason } from "./client";
 export { APPWIRE_PROTOCOL_VERSION, AppwireClient } from "./client";
 export type { AppwireClientLike } from "./clientLike";
+export type { InputAttachment } from "./composerInput";
+export { buildComposerInput, buildInput } from "./composerInput";
 export {
   firstLine,
   formatCharCount,
